# docs/blueprint/api-openapi.yaml
openapi: 3.0.3
info:
  title: SAFE API
  version: 0.1.0
servers:
  - url: https://api.safe.dz
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string, minLength: 6 }
                displayName: { type: string }
      responses:
        '200': { description: Token + user }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string }
      responses:
        '200': { description: Token + user }
  /listings:
    get:
      summary: Search listings
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
      responses:
        '200': { description: List of listings }
    post:
      summary: Create listing
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, price, category, city]
              properties:
                title: { type: string }
                description: { type: string }
                price: { type: integer }
                category: { type: string }
                city: { type: string }
                images: { type: array, items: { type: string } }
      responses:
        '200': { description: Listing }
  /reservations:
    get:
      summary: My reservations
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: List of reservations }
    post:
      summary: Create reservation
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [start, end, city]
              properties:
                listingId: { type: string }
                start: { type: string, format: date-time }
                end: { type: string, format: date-time }
                city: { type: string }
                notes: { type: string }
      responses:
        '200': { description: Reservation }
  /admin/users:
    get:
      summary: Admin users list
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: Users }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT